requirements.txt
numpy==1.26.4
chainlit==0.7.700   # 1.1.402 
openai>=1.26.0
pymupdf==1.24.9
pdfplumber==0.11.4
qdrant-client==1.11.2
langchain-text-splitters
langchain-core==0.2.27 
langchain-community==0.2.10
langchain-experimental==0.0.64 
langgraph-checkpoint==1.0.6 
langgraph==0.2.16 
langchain-qdrant==0.1.3
langchain-openai==0.1.9